# src/loaders/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Rest base repo.",
    "description": "## This is a base api to start new projects.",
    "version": "0.0.1",
    "contact": {
      "email": "adriansanchez88@gmail.com"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
  "schemes": "http",
  "definitions":{
    "User":{
      "type": "object",
      "properties": {
        "id":{
          "description": "The user id.",
          "type": "integer",          
          "example": "3"
        },
        "name":{
          "description": "The user name.",
          "type":"string",          
          "example":"Adrian"
        }
      },
      "required":["id"]
    }
  },
  "tags": [    
    {
      "name": "users",
      "description": "Operations about users."
    }
  ],
  "paths": {    
    "/users/": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get a list of all existing users.",
        "description": "Get a list of all existing users.",
        "operationId": "getAllUsers",
        "produces": [
          "application/json"
        ],        
        "responses": {
          "200": {
            "description": "The list of users.",
            "schema": {
              "$ref": "#/definitions/User"              
            }
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create a new user.",
        "description": "Create a new user.",
        "operationId": "createUser",
        "produces": [
          "application/json"
        ],        
        "responses": {
          "201": {
            "description": "The user created.",
            "schema": {
              "$ref": "#/definitions/User"              
            }
          }
        }
      }
    },
    "/users/{id}": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Update an existing user by providing the id.",
        "description": "Update an existing user by providing the id.",
        "operationId": "updateUser",
        "produces": [
          "application/json"
        ],      
        "parameters":[          
          {
            "name": "id",
            "in": "path",
            "description": "The id of the user that needs to be updated.",
            "required": true,
            "type": "integer"
          }          
        ], 
        "responses": {
          "200": {
            "description": "The user updated",
            "schema": {
              "$ref": "#/definitions/User"              
            }
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Delete an existing user by providing the id.",
        "description": "Delete an existing user by providing the id.",
        "operationId": "deleteUser",
        "produces": [
          "application/json"
        ],
        "parameters":[          
          {
            "name": "id",
            "in": "path",
            "description": "The id of the user that needs to be deleted.",
            "required": true,
            "type": "integer"
          }          
        ],      
        "responses": {
          "200": {
            "description": "The user deleted.",
            "schema": {
              "$ref": "#/definitions/User"              
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "InventoryItem": {
        "required": [
          "id",
          "manufacturer",
          "name",
          "releaseDate"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "d290f1ee-6c54-4b01-90e6-d701748f0851"
          },
          "name": {
            "type": "string",
            "example": "Widget Adapter"
          },
          "releaseDate": {
            "type": "string",
            "format": "date-time",
            "example": "2016-08-29T09:12:33.001Z"
          },
          "manufacturer": {
            "$ref": "#/components/schemas/Manufacturer"
          }
        }
      },
      "Manufacturer": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "ACME Corporation"
          },
          "homePage": {
            "type": "string",
            "format": "url",
            "example": "https://www.acme-corp.com"
          },
          "phone": {
            "type": "string",
            "example": "408-867-5309"
          }
        }
      }
    }
  }
}